requirements.txt
pandas
openai
python-dotenv
numpy
tqdm
scikit-learn